Added constants for astparser node types, kinds and usage kinds
