+++ modules/astparser/src/ast/ASTSymbolInfo.h
@@ -34,88 +34,8 @@
 
 #include "ast/ASTName.h"
 #include "ast/ASTNode.h"
-
-//! Describes the kind of symbol.
-enum ASTSymbolKind {
-    ASYK_None = 0,
-    ASYK_File = 1,
-    ASYK_Module = 2,
-    ASYK_Namespace = 3,
-    ASYK_Package = 4,
-    ASYK_Class = 5,
-    ASYK_Method = 6,
-    ASYK_Property = 7,
-    ASYK_Field = 8,
-    ASYK_Constructor = 9,
-    ASYK_Enum = 10,
-    ASYK_Interface = 11,
-    ASYK_Function = 12,
-    ASYK_Variable = 13,
-    ASYK_Constant = 14,
-    ASYK_String = 15,
-    ASYK_Number = 16,
-    ASYK_Boolean = 17,
-    ASYK_Array = 18,
-};
-
-//! Describes in what situation a symbol was used.
-enum ASTSymbolUsageKind {
-    ASUK_None = 0,
-
-    // Declarations.
-    ASUK_ClassDeclName = 100,
-    ASUK_ConstantDeclName = 101,
-    ASUK_FuncDeclName = 102,
-    ASUK_FuncReturnType = 103,
-    ASUK_NamespaceDeclName = 104,
-    ASUK_SuperclassDeclName = 105,
-    ASUK_VarDeclName = 106,
-    ASUK_VarDeclTypeName = 107,
-    ASUK_HashDeclName = 108,
-    ASUK_HashMemberName = 109,
-
-    // Expressions.
-    ASUK_AccessVariable = 200,
-    ASUK_AccessMember = 201,
-    ASUK_AssignmentLeft = 202,
-    ASUK_AssignmentRight = 203,
-    ASUK_BinaryLeft = 204,
-    ASUK_BinaryRight = 205,
-    ASUK_CallTarget = 206,
-    ASUK_CallArgs = 207,
-    ASUK_CaseExpr = 208,
-    ASUK_CastType = 209,
-    ASUK_CastObject = 210,
-    ASUK_FindData = 211,
-    ASUK_HashValue = 212,
-    ASUK_IndexVariable = 213,
-    ASUK_IndexIndex = 214,
-    ASUK_ListElement = 215,
-    ASUK_ReturnsType = 216,
-    ASUK_TernaryCond = 217,
-    ASUK_TernaryTrue = 218,
-    ASUK_TernaryFalse = 219,
-    ASUK_Unary = 220,
-
-    // Statements.
-    ASUK_ContextStmtName = 300,
-    ASUK_ContextStmtData = 301,
-    ASUK_DoWhileStmtCond = 302,
-    ASUK_ExprStmtExpr = 303,
-    ASUK_ForeachStmtSrc = 304,
-    ASUK_ForeachStmtVal = 305,
-    ASUK_ForStmtCond = 306,
-    ASUK_ForStmtInit = 307,
-    ASUK_ForStmtIter = 308,
-    ASUK_IfStmtCond = 309,
-    ASUK_ReturnStmtVal = 310,
-    ASUK_SummarizeStmtName = 311,
-    ASUK_SummarizeStmtData = 312,
-    ASUK_SwitchStmtVar = 313,
-    ASUK_ThrowStmtExpr = 314,
-    ASUK_TryStmtCatchVar = 315,
-    ASUK_WhileStmtCond = 316,
-};
+#include "ast/ASTSymbolKind.h"
+#include "ast/ASTSymbolUsageKind.h"
 
 class ASTSymbolInfo : public ASTNode {
 public:
